some tweaks to the seekbars in order to improve handling of touch of the extremities
I have tweaked touchEvents handling a little bit :
-I removed y from the checks, if the event if too far from the seekbar
on the vertical axis, it won't be dispatched to it anyway. its values
were a bit too narrow, creating false negatives.
-as long as the user touches the view, mIsMovingPointer = true.
Otherwise touches at the view extremities were not accepted because
they were not originating from the views, resulting in a pointer stuck
in an extremity.
-I simplified calculateColor a little bit. The behavior is still the
same, it is just more readable that way.