Changed selected color when bar pointers are moved
When event ACTION_UP is triggered the picker changes the color selected,
however it did not do so when the user changes the color using the bars.
The code was edited to notify the onColorSelectedListener to change the
selected color once the ACTION_UP event is triggered on a bar.